Added: - Major code refactoring by moving all the Xspress3 specific code into an AMXspress3XRFDetector class while keeping the BioXAS specific code in their classes. - Had to add another hack into the VESPERS XRF file loader to get the old XRF scans to load out of the database. #1192
